+++ macros/get_order_line_refund_columns.sql
@@ -2,7 +2,7 @@
 
 {% set columns = [
     {"name": "_fivetran_synced", "datatype": dbt.type_timestamp()},
-    {"name": "id", "datatype": dbt.type_numeric(), "alias": "order_line_refund_id"},
+    {"name": "id", "datatype": dbt.type_numeric()},
     {"name": "location_id", "datatype": dbt.type_numeric()},
     {"name": "order_line_id", "datatype": dbt.type_numeric()},
     {"name": "subtotal", "datatype": dbt.type_numeric()},