+++ macros/get_refund_columns.sql
@@ -3,7 +3,7 @@
 {% set columns = [
     {"name": "_fivetran_synced", "datatype": dbt.type_timestamp()},
     {"name": "created_at", "datatype": dbt.type_timestamp()},
-    {"name": "id", "datatype": dbt.type_numeric(), "alias": "refund_id"},
+    {"name": "id", "datatype": dbt.type_numeric()},
     {"name": "note", "datatype": dbt.type_string()},
     {"name": "order_id", "datatype": dbt.type_numeric()},
     {"name": "processed_at", "datatype": dbt.type_timestamp()},