+++ models/stg_shopify__order_adjustment.sql
@@ -23,7 +23,7 @@ fields as (
             union_database_variable='shopify_union_databases') 
         }}
         
-    from source
+    from base
 ),
 
 final as (