+++ models/stg_shopify__order_line.sql
@@ -21,7 +21,7 @@ fields as (
             union_database_variable='shopify_union_databases') 
         }}
 
-    from source
+    from base
 
 ),
 
@@ -72,4 +72,3 @@ final as (
 
 select * 
 from final
-