+++ models/stg_shopify__product.sql
@@ -1,10 +1,11 @@
-with source as (
+with base as (
 
-    select * from {{ ref('stg_shopify__product_tmp') }}
+    select * 
+    from {{ ref('stg_shopify__product_tmp') }}
 
 ),
 
-renamed as (
+fields as (
 
     select
     
@@ -15,21 +16,37 @@ renamed as (
             )
         }}
 
-      --The below script allows for pass through columns.
-      {% if var('product_pass_through_columns') %}
-      ,
-      {{ var('product_pass_through_columns') | join (", ")}}
-
-      {% endif %}
-
         {{ fivetran_utils.source_relation(
             union_schema_variable='shopify_union_schemas', 
             union_database_variable='shopify_union_databases') 
         }}
 
-    from source
+    from base
+
+),
+
+final as (
+    
+    select
+        created_at,
+        handle,
+        id as product_id,
+        product_type,
+        published_at,
+        published_scope,
+        title,
+        updated_at,
+        vendor,
+        status,
+        _fivetran_deleted as is_deleted,
+        _fivetran_synced,
+        source_relation
+
+        {{ fivetran_utils.fill_pass_through_columns('product_pass_through_columns') }}
+
+from fields
 
 )
 
-select * from renamed
-
+select * 
+from final